llm_api_311/requirements.txt
fastapi
uvicorn
pydantic
PyYAML
aiofiles
requests
qdrant-client
torch
transformers
httpx
huggingface-hub
pillow
torch
torchaudio
torchvision